fix code convention

+++ src/user/user.service.ts
@@ -3,61 +3,63 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { Repository } from 'typeorm';
-import { User } from './user.entity';
+import { UserEntity } from './user.entity';
 import { UserRequestDto } from './dto/user-request.dto';
 import { InjectRepository } from '@nestjs/typeorm';
-import { UserResponse } from './response/user.response';
-import { NotFoundError } from 'rxjs';
+import { User } from './user.domain';
 
 @Injectable()
 export class UserService {
   constructor(
-    @InjectRepository(User)
-    private readonly userRepository: Repository<User>,
-  ) {}
+    @InjectRepository(UserEntity)
+    private readonly userRepository: Repository<UserEntity>,
+  ) { }
 
-  async findAll(): Promise<UserResponse[]> {
+  async findAll(): Promise<User[]> {
     const users = await this.userRepository.find();
-    return users.map((user) =>
-      UserResponse.toUserResponse(user),
-    );
+
+    return users.map(User.fromEntities);
   }
 
-  async findOne(id: number): Promise<UserResponse | null> {
+  async findOne(id: number): Promise<User> {
     const user = await this.userRepository.findOneBy({
       id,
     });
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    return UserResponse.toUserResponse(user);
+    return User.fromEntities(user);
   }
 
   async create(
     userRequestDto: UserRequestDto,
-  ): Promise<UserResponse> {
-    const createUser =
-      this.userRepository.create(userRequestDto);
-    return UserResponse.toUserResponse(
-      await this.userRepository.save(createUser),
-    );
+  ): Promise<User> {
+    const createUser = this.userRepository.create(userRequestDto);
+
+    return User.fromEntities(await this.userRepository.save(createUser));
   }
 
   async update(
     id: number,
     userRequestDto: UserRequestDto,
-  ): Promise<UserResponse | null> {
+  ): Promise<User> {
     await this.userRepository.update(id, userRequestDto);
-    const user = await this.userRepository.findOneBy({
-      id,
-    });
+
+    const user = await this.userRepository.findOneBy({ id });
+
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    return UserResponse.toUserResponse(user);
+
+    return User.fromEntities(user);
   }
 
   async delete(id: number): Promise<void> {
+    const user = await this.userRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     await this.userRepository.delete(id);
   }
 }
